# Conflicts:
#	src/Controller/QuestController.php
#	templates/security/login.html.twig

+++ src/Controller/QuestController.php
@@ -13,6 +13,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Contracts\Service\Attribute\Required;
 
 #[Route('/quest', name: 'quest_')]
 final class  QuestController extends AbstractController
@@ -21,7 +22,7 @@ final class  QuestController extends AbstractController
     public function index(QuestRepository $questRepository): Response
     {
         return $this->render('quest/index.html.twig', [
-            'quests' => $questRepository->findAll(),
+            'quests' => $questRepository->findAll()
         ]);
     }
 
@@ -39,8 +40,8 @@ final class  QuestController extends AbstractController
         $quest = new Quest();
         if ($id != null) {
             $quest = $questRepository->find($id);
-            if ($quest->getUsers() != $this->getUser()) {
-                throw $this->createAccessDeniedException("Vas saboter la Quest d'autrui, malautru!");
+            if($quest->getPromoter() != $this->getUser()){
+                throw $this->createAccessDeniedException("Vas saboter la quête d'autrui, malautru!");
             }
         }
         $status = $statusRepository->findAll();
@@ -102,14 +103,26 @@ final class  QuestController extends AbstractController
         return $this->redirectToRoute('quest_index');
     }
 
-    #[Route('/{id}', name: 'delete', methods: ['POST'])]
-    public function delete(Request $request, Quest $quest, EntityManagerInterface $entityManager): Response
+
+    #[Route('/delete/{id}', name: 'delete', requirements:['id'=>'\d+'])]
+    public function delete(
+        Request $request,
+        Quest $quest,
+        EntityManagerInterface $entityManager,
+        QuestRepository $questRepository,
+        int $id): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $quest->getId(), $request->getPayload()->getString('_token'))) {
+        $quest = $questRepository->find($id);
+
+        if ($this->isCsrfTokenValid('delete'.$quest->getId(), $request->getPayload()->getString('_token'))) {
+
+            if($quest->getPromoter() != $this->getUser() && !$this->isGranted('ROLE_ADMIN')){
+                throw $this->createAccessDeniedException("Ne destroies point la sortie qui n'est nulle la tienne!");
+            }
+
             $entityManager->remove($quest);
             $entityManager->flush();
         }
-
-        return $this->redirectToRoute('quest_index', ['id' => $quest->getId()], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('quest_index', ['id'=>$quest->getId()], Response::HTTP_SEE_OTHER);
     }
 }
